Movie Seat Booking System
User can log in with valid email and password if they don't have then they can register themselves. User will Enter the movie name and timing and then will book the seats if the seats are already booked for that show then he can't book that particular seat. All records of every user is saved in database. Behind the scene, every user is given the unique user id and with that user id the seat booking is done.